requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
langchain
langchain-openai
langchain_community
langchain-huggingface
sentence-transformers
faiss-cpu
PyMuPDF
requests